+++ XMTPiOSExample/XMTPiOSExample/Views/LoggedInView.swift
@@ -1,0 +1,17 @@
+//
+//  LoggedInView.swift
+//  XMTPiOSExample
+//
+//  Created by Pat Nakajima on 11/22/22.
+//
+
+import SwiftUI
+import XMTP
+
+struct LoggedInView: View {
+	var wallet: XMTP.Wallet
+
+	var body: some View {
+		Text("We're in")
+	}
+}
